+++ tests/dropout_test.py
@@ -13,7 +13,7 @@ def test_dropout(keys):
         m = Dropout(p=0.75)
         return m(rng, x, is_training)
 
-    x = IrrepsArray.from_array(irreps, irreps.randn(next(keys), (-1,)))
+    x = IrrepsArray(irreps, irreps.randn(next(keys), (-1,)))
     params = b.init(next(keys), next(keys), x)
 
     y = b.apply(params, next(keys), x, is_training=False)
@@ -23,7 +23,7 @@ def test_dropout(keys):
     assert ((y.array == (x.array / 0.25)) | (y.array == 0)).all()
 
     def wrap(x):
-        x = IrrepsArray.from_array(irreps, x)
+        x = IrrepsArray(irreps, x)
         return b.apply(params, keys[0], x).array
 
     assert_equivariant(wrap, rng_key=next(keys), args_in=[x.array], irreps_in=[irreps], irreps_out=[irreps])